Implement book service with CRUD operations
A book service has been implemented with Create, Read, Update, and Delete (CRUD) operations. New files have been created for the service implementation, and Lombok library has been added to the build.gradle file. The service is supported by a new Book repository and model which are also included in this commit. A new BookController file is created for the REST endpoints to manage books.